Build 0.1

+++ course/src/main/java/com/educandoweb/course/services/UserService.java
@@ -3,11 +3,17 @@ package com.educandoweb.course.services;
 import java.util.List;
 import java.util.Optional;
 
+import org.hibernate.action.internal.EntityActionVetoException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.educandoweb.course.entities.UserTable;
+import com.educandoweb.course.exceptions.DataBaseException;
+import com.educandoweb.course.exceptions.ResourceNotFoundException;
 import com.educandoweb.course.repositories.UserRepository;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class UserService {
@@ -21,7 +27,7 @@ public class UserService {
 	
 	public UserTable findById(Long id){
 		Optional<UserTable> obj = repository.findById(id);
-		return obj.get();
+		return obj.orElseThrow(() -> new ResourceNotFoundException(id));
 	}
 	
 	public UserTable insert(UserTable obj) {
@@ -29,17 +35,32 @@ public class UserService {
 	}
 	
 	public void delete(Long id) {
-		repository.deleteById(id);
+		try {
+			
+			if (repository.existsById(id)) {
+	            repository.deleteById(id);			
+	        } else {				
+	            throw new ResourceNotFoundException(id);			
+	        }	
+		  } catch (DataIntegrityViolationException e) {			
+		        throw new DataBaseException(e.getMessage());	
+		}
+		
 	}
 	
 	public UserTable update(Long id, UserTable obj) {
-		UserTable entity = repository.getReferenceById(id);
-		updateData(entity, obj);
-		return repository.save(entity);
+		try {
+			UserTable entity = repository.getReferenceById(id);
+			updateData(entity, obj);
+			return repository.save(entity);	
+		} catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundException(id);	
+		}
 	}
-	public void updateData(UserTable entity, UserTable obj) {
+
+	private void updateData(UserTable entity, UserTable obj) {
 		entity.setName(obj.getName());
 		entity.setEmail(obj.getEmail());
 		entity.setPhone(obj.getPhone());
-	}	
+	}
 }
